feat(oidc): claims parameter support
This adds support for the OpenID Connect 1.0 claims parameter. Consequently, this means clients will have to request specific claims they want in the ID Token which are not part of the standard claims and the scoped claims will be by default returned instead by the UserInfo endpoint more closely aligning with the specification.

Signed-off-by: James Elliott <james-d-elliott@users.noreply.github.com>