fix(web): adjust router basename behaviour
This change adjusts the behaviour if Authelia is run with a sub-path and is visited without a trailing slash on the specified sub-path.
In 4.37.5 the base path would get normalized without a trailing slash, however, would cause issues when a refresh was completed while carrying a redirection query string.
In 4.38.x this was changed so the sub-path would not be normalized without the trailing slash and that it was therefore necessary. This change in behaviour could be observed as a regression by users with learned behaviours.

Signed-off-by: Amir Zarrinkafsh <nightah@me.com>